+++ cli/cli/commands/enclave/enclave.go
@@ -26,7 +26,7 @@ func init() {
 	EnclaveCmd.AddCommand(ls.EnclaveLsCmd.MustGetCobraCommand())
 	EnclaveCmd.AddCommand(inspect.EnclaveInspectCmd.MustGetCobraCommand())
 	EnclaveCmd.AddCommand(new.NewCmd)
-	EnclaveCmd.AddCommand(stop.StopCmd)
+	EnclaveCmd.AddCommand(stop.EnclaveStopCmd.MustGetCobraCommand())
 	EnclaveCmd.AddCommand(rm.EnclaveRmCmd.MustGetCobraCommand())
 	EnclaveCmd.AddCommand(dump.EnclaveDumpCmd.MustGetCobraCommand())
 }