+++ cli/cli/commands/enclave/new/new.go
@@ -3,7 +3,6 @@ package new
 import (
 	"context"
 	"fmt"
-	"github.com/kurtosis-tech/container-engine-lib/lib/backend_impls/docker/backend_creator"
 	"github.com/kurtosis-tech/kurtosis-cli/cli/command_str_consts"
 	"github.com/kurtosis-tech/kurtosis-cli/cli/defaults"
 	"github.com/kurtosis-tech/kurtosis-cli/cli/helpers/engine_manager"
@@ -26,6 +25,9 @@ const (
 
 	// Signifies that an enclave ID should be auto-generated
 	autogenerateEnclaveIdKeyword = ""
+
+	// TODO DElete when this comes from disk
+	clusterName = "docker"
 )
 
 var apiContainerVersion string
@@ -78,13 +80,11 @@ func run(cmd *cobra.Command, args []string) error {
 
 	ctx := context.Background()
 
-	// TODO REFACTOR: we should get this backend from the config!!
-	var apiContainerModeArgs *backend_creator.APIContainerModeArgs = nil  // Not an API container
-	kurtosisBackend, err := backend_creator.GetLocalDockerKurtosisBackend(apiContainerModeArgs)
+
+	engineManager, err := engine_manager.NewEngineManager(clusterName)
 	if err != nil {
-		return stacktrace.Propagate(err, "An error occurred getting a Kurtosis backend connected to local Docker")
+		return stacktrace.Propagate(err, "An error occurred creating an engine manager using cluster '%v'", clusterName)
 	}
-	engineManager := engine_manager.NewEngineManager(kurtosisBackend)
 	engineClient, closeClientFunc, err := engineManager.StartEngineIdempotentlyWithDefaultVersion(ctx, defaults.DefaultEngineLogLevel)
 	if err != nil {
 		return stacktrace.Propagate(err, "An error occurred creating a new Kurtosis engine client")